refactor(ts-client): consolidate to single minimal example
Replace tutorial/main.ts with a simple ERC-20 swap example using
the new API (swapPayload, swapSigningHash, assembleSignedSwap,
approvalSigningHash). Remove the extra swap_erc20.ts. Update the
quickstart docs to match.
